requirements-seo.txt
# Dependencies for scripts/generate-seo-assets.py
# Install with: pip install -r requirements-seo.txt
Pillow>=10.0.0
